python_worker/requirements.txt
requests
redis
boto3
python-dotenv
pyaml
asyncio
